Reduce precision of node version
The version of node installed on the EC2 instances is only specified at
the major version number. This means it will be the latest v20 version
at the time that the infrastructure was built. This means fixing node at
the patch level in .nvmrc makes less sense.

Also added ability to run unit tests with generating coverage